custom CKRecordID based on the existing objectId
